Create run.py
process the .txt files (named 001.txt, 002.txt, ...) in the './supplier-data/descriptions/' directory and save them in a data structure and upload them via JSON. 
all files are written in the following format, with each piece of information on its own line:
name
weight (in lbs)
description

the data model in the Django application fruit has the following fields: name, weight, description and image_name. 
the weight field is defined as an integer field. so process the weight information of the fruit from the .txt file, need to convert it into an integer. 

the image_name field will allow the system to find the image associated with the fruit.